emailModel renamed, moved and updated
